lobby ui for the lobby list
used the ui groups for the first time.
The lobby list has title, section titles, a box that shows the current lobbies, their name, player count, and button to join, and an input and button to join private lobby. there is a button to go back, refresh the list, and create your own lobby.
so far this is all just set out and doesnt have any functionality